migrate qvm system to be entirely C-based

+++ src/game_jk2sp.cpp
@@ -343,7 +343,7 @@ intptr_t JK2SP_syscall(intptr_t cmd, ...) {
         char* buffer = (char*)args[0];
         intptr_t bufferSize = args[1];
 
-        strncpyz(buffer, s_entity_tokens[s_tokencount++].c_str(), bufferSize);
+        strncpyz(buffer, s_entity_tokens[s_tokencount++].c_str(), (size_t)bufferSize);
         ret = qtrue;
         break;
     }
